output window: click on overview item adjusts output view (jump)
On MacOS clicking on the overview item did not result in a correct
adjustment of the output window. The reason is that view->y is of type
integer but type long is stored in the treestore. This requires correct
type casting.

+++ src/ui/gui/psppire-output-view.c
@@ -375,7 +375,7 @@ psppire_output_view_put (struct psppire_output_view *view,
       gtk_tree_store_set (store, &iter,
                           COL_NAME, ds_cstr (&name),
 			  COL_ADDR, item,
-                          COL_Y, view->y,
+                          COL_Y, (glong)(view->y),
                           -1);
       ds_destroy (&name);
